requirements.txt
argparse==1.4.0
cmake
dlib==19.24.1
imutils==0.5.4
numpy==1.24.3
opencv-python==4.5.5.64
requests
schedule==1.1.0
scipy==1.10.1
cmake==3.22.5
requests
Flask[async]==2.3.2
sounddevice
spotipy==2.23.0
python-dotenv==1.0.0
pytradfri==13.0.0
pydantic==1.10.9